Include clean up step
