+++ app/models/data_importer.rb
@@ -2,7 +2,7 @@ require 'json'
 require 'httparty'
 require 'ap'
 require_relative './hash'
-require_relative './data_saver'
+require_relative './file_generator'
 
 class DataImporter
 
@@ -11,8 +11,9 @@ class DataImporter
   def self.get_sampling(resource_id, table_name, limit=25)
     response = HTTParty.get("#{TLD}/datastore_search\?resource_id\=#{resource_id}&limit=#{limit}")
     result = response['result']
-    ap result
-    DataSaver.save_sample_data(result, table_name)
+    # ap result
+    # DataSaver.save_sample_data(result, table_name)
+    FileGenerator.generate_initial(resource_id, table_name, result)
   end
 
   def self.get_all(resource_id)
