+++ app/models/data_saver.rb
@@ -6,13 +6,6 @@ require_relative './file_generator'
 Dir["./*.rb"].each {|file| require file }
 
 class DataSaver
-  # As more datatypes are encountered, add to this hash
-  FIELDS_MAP = {
-    "int4": "integer",
-    "numeric": "integer",
-    "text": "string",
-    "timestamp": "datetime"
-  }
 
   def self.save_all(resource_id, table_name)
     FileGenerator.generate_resource(resource_id, table_name)
@@ -20,33 +13,15 @@ class DataSaver
   end
 
   def self.save_resources(resource_id, table_name)
-        load File.join( File.dirname(__FILE__), "#{table_name.singularize}.rb")
+    load File.join( File.dirname(__FILE__), "#{table_name.singularize}.rb")
     class_name = table_name.classify.constantize
     results = DataImporter.get_all(resource_id)
-
+    # binding.pry
     results.each do |result|
+      result = class_name.clean_data(result)
       class_name.create(result)
     end
     nil
   end
-
-  def self.save_sample_data(result, table_name)
-    FileGenerator.generate_json(result)
-
-    attributes = self.parse_attributes(result['fields'])
-    FileGenerator.generate_migration(table_name, attributes)
-  end
-
-  private
-
-    def self.parse_attributes(fields)
-      attributes = {}
-      fields.each do |field|
-        attr_name = field["id"]
-        data_type = FIELDS_MAP[field["type"].to_sym]
-        attributes[attr_name] = data_type
-      end
-      attributes
-    end
 
 end
